backend/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-MySQLdb==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.1
PyYAML==6.0
Werkzeug==2.1.2
zipp==3.8.1